goit-ds-03/pyproject.toml
[tool.poetry]
name = "goit-ds-03"
version = "0.1.0"
description = ""
authors = ["Siracenco Serghei <siracencoserghei@email.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymongo = "^4.12.1"
mongoengine = "^0.29.1"
python-dotenv = "^1.1.0"
docker = "^7.1.0"
faker = "^37.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
